src/data_loader/requirements.txt
pandas~=2.3.1
psycopg2-binary~=2.9.10
python-dotenv~=1.1.1